Change pushup_window display state from fullscreen to maximized
In the old method, the window expanded to occupy the full screen, and the '-topmost' attribute ensured that it is displayed above other windows. However, one issue with this method is that while the window is displayed above other windows, the focus is not on the window. This creates a problem while enabling a keypress event that fires the pushup_timer. Key presses that are meant for this window impact the last open window. Instead, if the window is maximized rather than changing the state to fullscreen, the deiconify() method that is used to maximize the window automatically brings focus as required.

- Also, set focus and bind enter keypress event to btn_start_pushup_timer